rename ProfileForm to EditProfile and ProfileRelatedForm to ProfileForm
