+++ src/components/ProfileForm/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "ProfileForm.js",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./ProfileForm.js"
+}
