--- src/components/ProfileRelatedForm/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "ProfileRelatedForm.js",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./ProfileRelatedForm.js"
-}
