+++ src/screens/NewProfile/NewProfile.js
@@ -32,8 +32,8 @@ import ProfileImage from 'components/ProfileImage';
 import { validateUserDetailsAction, registerOnBackendAction } from 'actions/onboardingActions';
 import { USERNAME_EXISTS, USERNAME_OK, CHECKING_USERNAME, INVALID_USERNAME } from 'constants/walletConstants';
 import { fontSizes, fontWeights } from 'utils/variables';
-import { InputTemplate, Form } from 'components/ProfileRelatedForm';
-import { Username, MAX_USERNAME_LENGTH } from 'components/ProfileRelatedForm/profileFormDefs';
+import { InputTemplate, Form } from 'components/ProfileForm';
+import { Username, MAX_USERNAME_LENGTH } from 'components/ProfileForm/profileFormDefs';
 
 const IntroParagraph = styled(Paragraph)`
   margin: 10px 0 50px;
